Attempt to fix the multiplayer issue
